Add shop log

+++ app/Http/Controllers/ItemController.php
@@ -3,16 +3,21 @@
 namespace App\Http\Controllers;
 
 use App\Models\Item;
+use App\Models\ShoppingLog;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Log;
 
 class ItemController extends Controller
 {
     public function index()
     {
-        $items = Item::where('deleteItem', 0)->get();
-
-        return view('items.index', compact('items'));
+        if(Auth::check())
+        {
+            $items = Item::where('deleteItem', 0)->get();
+            return view('items.index', compact('items'));
+        };
+        return redirect()->route('login');
     }
 
     public function IndexApi()
@@ -20,48 +25,69 @@ class ItemController extends Controller
         $items = Item::where('deleteItem', 0)->get();
 
         echo json_encode($items);
+
     }
 
     public function editForm($id)
     {
-        $item = Item::findOrFail($id);
-        return view('items.edit_form', compact('item'));
+        if(Auth::check())
+        {
+            $item = Item::findOrFail($id);
+            return view('items.edit_form', compact('item'));
+        };
+        return redirect()->route('login');
     }
 
     public function update(Request $request, $id)
     {
+        $user = Auth::user();
+
         $item = Item::findOrFail($id);
         $item->quantity = $request->input('quantity');
         $item->measure = $request->input('measure');
         $item->Items = $request->input('Items');
+        $item->user_id = $user->id;
         $item->save();
+
+        $this->logAction('edit', $item->id);
 
         return redirect()->route('items.index')->with('success', 'Item updated successfully!');
     }
 
     public function createForm()
     {
-        return view('items.create_form');
+        if(Auth::check())
+        {
+            return view('items.create_form');
+        };
+        return redirect()->route('login');
     }
 
     public function store(Request $request)
     {
+        $user = Auth::user();
+
         $item = new Item;
         $item->quantity = $request->input('quantity');
         $item->measure = $request->input('measure');
         $item->Items = $request->input('Items');
+        $item->insertDate = date('Y-m-d H:i:s',time());
+        $item->user_id = $user->id;
         $item->save();
 
+        $this->logAction('create', $item->id);
         return redirect()->route('items.index')->with('success', 'Item created successfully!');
     }
 
     public function toggleNakupljeno(Request $request, $id)
     {
         try {
+            $user = Auth::user();
             $item = Item::findOrFail($id);
             $item->nakupljeno = $request->input('nakupljeno');
+            $item->user_id = $user->id;
             $item->save();
-
+            $this->logAction('buy', $item->id);
             return response()->json(['success' => true]);
         } catch (\Exception $e) {
             // Log the exception
@@ -74,11 +100,14 @@ class ItemController extends Controller
 
     public function toggleNakupljenoAjax(Request $request, $id)
     {
-        // Logic to toggle the 'nakupljeno' field for a specific item
+        $user = Auth::user();
         $item = Item::findOrFail($id);
         $item->nakupljeno = !$item->nakupljeno;
         $item->buyDate = (($item->nakupljeno == 1)?date('Y-m-d H:i:s',time()):$item->buyDate);
+        $item->user_id = $user->id;
         $item->save();
+
+        $this->logAction('buy', $item->id);
 
         $response = ($item->nakupljeno)?1:0;
         return response()->json(['success' => $response]);
@@ -92,11 +121,31 @@ class ItemController extends Controller
      */
     public function destroy(Item $item,$id)
     {
-        $item = Item::findOrFail($id);
-        $item->deleteItem = 1;
-        $item->deleteDay = date('Y-m-d H:i:s',time());
-        $item->save();
+        if(Auth::check())
+        {
+            $user = Auth::user();
+            $item = Item::findOrFail($id);
+            $item->deleteItem = 1;
+            $item->deleteDay = date('Y-m-d H:i:s',time());
+            $item->user_id = $user->id;
+            $item->save();
 
-        return response()->json(['message' => 'Item deleted successfully']);
+            $this->logAction('delete', $item->id);
+
+            return response()->json(['message' => 'Item deleted successfully']);
+        };
+        return redirect()->route('login');
     }
+
+    private function logAction($action, $item_id)
+{
+    // Log the action in shopping_logs table
+    $user = Auth::user(); // Get the currently authenticated user
+
+    ShoppingLog::create([
+        'item_id' => $item_id,
+        'action' => $action,
+        'user_id' => $user->id,
+    ]);
+}
 }
